requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.3
python==3.11.4